refactor: split TaskGroupWithTimeout into its own package
